Uses more explicit var names in v3 keygen
The per-service vars weren't intuitively explained in the JSON file.
Updated the var names so that their use in the corresponding Ansible
logic is more readable, and less error-prone.

We should move the v3 keygen logic out of securedrop-admin and into
Ansible, since otherwise it will not work in developer environments, or
CI.