Formats v3 client auth files in restrict-direct-access
Implementing suggestion by @zenmonkeykstop, to format the files directly
on the Admin Workstation in the format required for ingestion by the
`tailsconfig` logic in securedrop-admin.

Modified the filename for the journalistv3 service, so that the existing
"if filename ends with '-aths'" logic will work, as with the SSH
services.

h/t to @redshiftzero for an assist on the Jinja.