Updates v3 fail-fast logic for admins
We're aiming to fail early if an Admin runs the install action
against a server with v3 configured, but without having run
sdconfig to enable v3 (indicates multiple Admins).

Updated the monstrous filter line to pull out the booleans, then inspect
that list in an assert statement. Seems to be working as expected, will
test with multiple configuration paths prior to squashing.

Better support for "staging" scenario, which has v3_onion_services=true,
but enable_ssh_over_tor=false, so mon-staging was failing the check.
Updated to accommodate.

If no onions exist, don't fail: we're about to set them up!